+++ src/sdk/OperationTracker.ts
@@ -1,6 +1,6 @@
 import axios from 'axios';
 
-import { Network, TransactionLinker, SimplifiedStatuses, StatusesResponse } from '../structs/Struct';
+import { Network, TransactionLinker, SimplifiedStatuses, StatusesResponse, StatusByOperationId } from '../structs/Struct';
 import { MAINNET_PUBLIC_LITE_SEQUENCER_ENDPOINTS, TESTNET_PUBLIC_LITE_SEQUENCER_ENDPOINTS } from './Consts';
 import { operationFetchError, statusFetchError } from '../errors';
 
@@ -40,7 +40,7 @@ export class OperationTracker {
         throw operationFetchError;
     }
 
-    async getOperationStatus(operationId: string): Promise<string> {
+    async getOperationStatus(operationId: string): Promise<StatusByOperationId> {
         for (const endpoint of this.customLiteSequencerEndpoints) {
             try {
                 const response = await axios.post<StatusesResponse>(`${endpoint}/status`, {
@@ -48,11 +48,11 @@ export class OperationTracker {
                 });
 
                 const result = response.data.response.find((s) => s.operation_id === operationId);
-                if (result?.error_message) {
-                    throw statusFetchError(result.error_message);
+                if (!result) {
+                    throw statusFetchError('operation is not found in response');
                 }
 
-                return result?.status || '';
+                return result;
             } catch (error) {
                 console.error(`Error fetching status transaction with ${endpoint}:`, error);
             }
@@ -69,7 +69,10 @@ export class OperationTracker {
             return SimplifiedStatuses.OperationIdNotFound;
         }
 
-        const status = await this.getOperationStatus(operationId);
+        const { status, error_message } = await this.getOperationStatus(operationId);
+        if (error_message) {
+            return SimplifiedStatuses.Failed;
+        }
         const finalStatus = isBridgeOperation ? this.BRIDGE_TERMINATED_STATUS : this.TERMINATED_STATUS;
         if (status == finalStatus) {
             return SimplifiedStatuses.Successful;